# ui/packages/tidb-dashboard-client/swagger/spec.json
{
    "swagger": "2.0",
    "info": {
        "title": "Dashboard API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/dashboard/api",
    "paths": {
...
        "/host/all": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Get information of all hosts",
                "operationId": "clusterInfoGetHostsInfo",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/clusterinfo.GetHostsInfoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/host/statistics": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Get cluster statistics",
                "operationId": "clusterInfoGetStatistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/clusterinfo.ClusterStatistics"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/topology/alertmanager": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Get AlertManager instance",
                "operationId": "getAlertManagerTopology",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/topology.AlertManagerInfo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/topology/alertmanager/{address}/count": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Get current alert count from AlertManager",
                "operationId": "getAlertManagerCounts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ip:port",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/topology/grafana": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Get Grafana instance",
                "operationId": "getGrafanaTopology",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/topology.GrafanaInfo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/topology/pd": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Get all PD instances",
                "operationId": "getPDTopology",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/topology.PDInfo"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/topology/scheduling": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Get all Scheduling instances",
                "operationId": "getSchedulingTopology",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/topology.SchedulingInfo"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/topology/store": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Get all TiKV / TiFlash instances",
                "operationId": "getStoreTopology",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/clusterinfo.StoreTopologyResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/topology/store_location": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Get location labels of all TiKV / TiFlash instances",
                "operationId": "getStoreLocationTopology",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/topology.StoreLocation"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/topology/ticdc": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Get all TiCDC instances",
                "operationId": "getTiCDCTopology",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/topology.TiCDCInfo"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/topology/tidb": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Get all TiDB instances",
                "operationId": "getTiDBTopology",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/topology.TiDBInfo"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/topology/tidb/{address}": {
            "delete": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Hide a TiDB instance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ip:port",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete ok"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/topology/tiproxy": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Get all TiProxy instances",
                "operationId": "getTiProxyTopology",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/topology.TiProxyInfo"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/topology/tso": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Get all TSO instances",
                "operationId": "getTSOTopology",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/topology.TSOInfo"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "clusterinfo.ClusterStatistics": {
            "type": "object",
            "properties": {
                "probe_failure_hosts": {
                    "type": "integer"
                },
                "stats_by_instance_kind": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/clusterinfo.ClusterStatisticsPartial"
                    }
                },
                "total_stats": {
                    "$ref": "#/definitions/clusterinfo.ClusterStatisticsPartial"
                },
                "versions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "clusterinfo.ClusterStatisticsPartial": {
            "type": "object",
            "properties": {
                "number_of_hosts": {
                    "type": "integer"
                },
                "number_of_instances": {
                    "type": "integer"
                },
                "total_logical_cores": {
                    "type": "integer"
                },
                "total_memory_capacity_bytes": {
                    "type": "integer"
                },
                "total_physical_cores": {
                    "type": "integer"
                }
            }
        },
        "clusterinfo.GetHostsInfoResponse": {
            "type": "object",
            "properties": {
                "hosts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/hostinfo.Info"
                    }
                },
                "warning": {
                    "$ref": "#/definitions/rest.ErrorResponse"
                }
            }
        },
        "clusterinfo.StoreTopologyResponse": {
            "type": "object",
            "properties": {
                "tiflash": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/topology.StoreInfo"
                    }
                },
                "tikv": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/topology.StoreInfo"
                    }
                }
            }
        },
...
        "hostinfo.CPUInfo": {
            "type": "object",
            "properties": {
                "arch": {
                    "type": "string"
                },
                "logical_cores": {
                    "type": "integer"
                },
                "physical_cores": {
                    "type": "integer"
                }
            }
        },
        "hostinfo.CPUUsageInfo": {
            "type": "object",
            "properties": {
                "idle": {
                    "type": "number"
                },
                "system": {
                    "type": "number"
                }
            }
        },
        "hostinfo.Info": {
            "type": "object",
            "properties": {
                "cpu_info": {
                    "$ref": "#/definitions/hostinfo.CPUInfo"
                },
                "cpu_usage": {
                    "$ref": "#/definitions/hostinfo.CPUUsageInfo"
                },
                "host": {
                    "type": "string"
                },
                "instances": {
                    "description": "Instances in the current host. The key is instance address",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/hostinfo.InstanceInfo"
                    }
                },
                "memory_usage": {
                    "$ref": "#/definitions/hostinfo.MemoryUsageInfo"
                },
                "partitions": {
                    "description": "Containing unused partitions. The key is path in lower case.\nNote: deviceName is not used as the key, since TiDB and TiKV may return different deviceName for the same device.",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/hostinfo.PartitionInfo"
                    }
                }
            }
        },
        "hostinfo.InstanceInfo": {
            "type": "object",
            "properties": {
                "partition_path_lower": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "hostinfo.MemoryUsageInfo": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "used": {
                    "type": "integer"
                }
            }
        },
        "hostinfo.PartitionInfo": {
            "type": "object",
            "properties": {
                "free": {
                    "type": "integer"
                },
                "fstype": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "rest.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "boolean"
                },
                "full_text": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "topology.AlertManagerInfo": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                }
            }
        },
        "topology.GrafanaInfo": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                }
            }
        },
        "topology.PDInfo": {
            "type": "object",
            "properties": {
                "deploy_path": {
                    "type": "string"
                },
                "git_hash": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "start_timestamp": {
                    "description": "Ts = 0 means unknown",
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "topology.SchedulingInfo": {
            "type": "object",
            "properties": {
                "deploy_path": {
                    "type": "string"
                },
                "git_hash": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "start_timestamp": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "topology.StoreInfo": {
            "type": "object",
            "properties": {
                "deploy_path": {
                    "type": "string"
                },
                "git_hash": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "port": {
                    "type": "integer"
                },
                "start_timestamp": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "status_port": {
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "topology.StoreLabels": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "topology.StoreLocation": {
            "type": "object",
            "properties": {
                "location_labels": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/topology.StoreLabels"
                    }
                }
            }
        },
        "topology.TSOInfo": {
            "type": "object",
            "properties": {
                "deploy_path": {
                    "type": "string"
                },
                "git_hash": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "start_timestamp": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "topology.TiCDCInfo": {
            "type": "object",
            "properties": {
                "cluster_name": {
                    "type": "string"
                },
                "deploy_path": {
                    "type": "string"
                },
                "git_hash": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "start_timestamp": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "status_port": {
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "topology.TiDBInfo": {
            "type": "object",
            "properties": {
                "deploy_path": {
                    "type": "string"
                },
                "git_hash": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "start_timestamp": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "status_port": {
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "topology.TiProxyInfo": {
            "type": "object",
            "properties": {
                "deploy_path": {
                    "type": "string"
                },
                "git_hash": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "start_timestamp": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "status_port": {
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JwtAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}